Fix double free attempt in OsGetCurrentDirectory.
If the allocation loop exits due to reaching the retry limit, the status
will be STATUS_BUFFER_TOO_SMALL. This causes CurrentDirectory to be freed
at the end of the routine, but CurrentDirectory would have already been
freed at the end of the loop. The routine may also have freed NULL if any
of the allocations of CurrentDirectory failed.

Thanks to d-hoke for finding this.