Clean up the enabling of checksum offloading.
This change breaks the RTL81xxx and DesignWare Ethernet device capabilities
into two separate fields: one for the supported capabilities and another
for the currently enabled capabilities. This allows the get/set information
API to make sure a caller isn't trying to set an unsupported capability.

This change also better protects the capability updates by adding a lock
for capability/configuration changes. The receive lock used to protect
this, but capability changes should not hold up the receive path. And for
DesignWare, the checkum offloading bit shares a register with link speed
configuration bits, so the lock is also acquired in DwepCheckLink to
synchronize register access.