Fixed a leftover mapping bug.
The Raspberry Pi hit an assert where MmpPageInAnonymousSection found a
mapping, but it wasn't a "present" mapping. We thing this is due to a bug
in the tearing down of the mapping that was there before this image
section.

The teardown code in MmpDestroyImageSectionMappings first marks all pages
as not-present, then works on freeing the physical pages. Finally, it
unmaps all pages. The problem is the first two of those steps started from
MinTouched and went until MaxTouched. The final unmapping was starting at
the section's base virtual address, which unmaps a different segment than
was worked on. This explains what we saw, as the non-present mapping had a
free physical address in it.

Also fixed a bug in isolate where a writable section should be
present/readable too. This was found by inspection, it didn't hit
anywhere.