Fixed a debugger deadlock.
A deadlock can arise with one process trying to queue a tracer child
signal onto its debugger, but the debugger has its own process lock
because it is dying, and trying to queue a Continue command onto the
child (so it's waiting for the CommandComplete event). For commands like
continue, don't wait on the Command Complete event, have faith that the
process will continue.

There is still an ugly problem where a child signal transformed into
something else after a process has ended will result in the process lock
being double-acquired, which results in an assert and a hang. The real
solution is to make it so that the signal code always owns the
SIGNAL_QUEUE_ENTRY allocations. Then the child signal callback can be
removed, and the child process can be reaped by whoever waits for it.