Add IP_MULTICAST_TTL support.
By default, outgoing multicast packets have a time-to-live of 1. They are
not typically meant to teach beyond the local network. This change adds
support for a socket option that modifies the default for a given socket.
This does not affect unicast packets, whose time-to-live values are still
dictated by the IP_TTL socket option (default of 63).