Add implicit join to IGMP all systems group.
Each IGMP host is supposed to join the all systems group (224.0.0.1) on
initialization. Rather than spin up the IGMP resources on link creation,
this change adds the all systems group when the first user request to
join a multicast group on that link arrives. This is also necessary for
now because only promiscuous mode is supported and not real multicast
filtering. An initial join on all machines would cause every Minoca system
to run in promiscuous mode.

This change also prevents any report or leave messages from being sent for
the all systems group. They aren't supposed to be sent as the membership
is implied.