Add USB mouse and HID drivers.
This change adds support for USB mice. Since not all mice appeared to
support the boot protocol (like the Qemu USB mouse), and the boot protocol
doesn't have support for the scroll wheel, I added a HID driver as well.
This driver is just a HID report parsing library, which the USB mouse
driver uses to figure out the report format and read out the values it
cares about. The HID driver doesn't serve any devices on its own.