Remove time zone support from the kernel.
The kernel only really had time zone support to support the RTC being set
in local time. Support for that is tricky to get right, and drags the
kernel into the policy mess of what time zone the system is set for.

This change removes all time zone data from the kernel, including loading
it in the boot loader, and managing it in the kernel. The kernel only sees
UTC time, user mode is now entirely responsible for viewing it through the
lens of a time zone.

This also means that the RTC is always set in UTC time.