Bug fixes for ImpElfPerformLibraryPathSubstitutions.
1. When a new buffer is allocated during the substitution, the old buffer
is freed but not replaced by the new buffer. This can result in use of
released memory and cause the caller to fail.

2. Variable size calculations were incorrect, causing segmentation faults.

3. The search for the next variable could begin in the middle of the
replacement string when it should start after it. This prevents processing
substitutions within substitutions, which is necessary to guard against an
infinite loop of substitutions.

4. The initial capacity of the path was not supplied properly by the
caller.

5. The loop that shifts the original string down when the replacement is
smaller than the variable started with the wrong index.

In general, the variable naming in this routine was confusing. This cleans
that up, hopefully making it easier to maintain.

These were documented by d-hoke in GitHub issue #112.