Ignore DT_RPATH if DT_RUNPATH is present.
When traversing up the chain of imports, a image's DT_RPATH is supposed to
be ignored if it also has a DT_RUNPATH. But don't abort the loop; just
keep going up until the chain is complete.

This also makes sure that the image associated with the path list is
passed to the helper routine, rather than the image trying to load the new
library. This will ensure that the correct substition happens for $ORIGIN
when that TODO is complete.