Add a routine to get the current directory.
Previously, a current directory path could only be queried by user mode
via IoSysGetCurrentDirectory. A kernel component in a forthcoming change
needs to get the current directory. This change provides that
functionality while sharing as much code as possible with the get current
directory system call.

As we are slowly changing buffer sizes from ULONG to UINTN, this change
got a little bigger, as I needed to fan out and change some of those.