Store the absolute path in the process environment.
When executing an image, convert relative paths into absolute paths by
getting the process's current directory and then save the absolute path in
the process environment. This is necessary to allow the image loader to
properly resolve $ORIGIN run paths. If the image loader only new relative
paths, a chdir function call could break a subsequent dlopen attempt.
In this bad scenario, $ORIGIN would be replaced by path that is relative
to the working directory before the chdir. The library search would fail.