Fix for spurious assert in DwarfpReadLeb128.
DWARF LEB128 and SLEB128s encode 7 bits at a time. I had an assert in
there to catch if values > 64bits were being encoded, because I'm using a
64-bit integer type to store those values (I'd rather not do
multiprecision math everywhere if I don't have to). The assert wasn't
quite correct because I forgot that the accumulated shift would be the
next rounded up factor of 7. So to fully encode 64 bits you'll end up with
a shift of 70 bits. This change relaxes the asserts to account for that.

d-hoke points out that the assert is no longer watertight, as encoded bits
between 65-70 will be silently truncated but not trigger the assert. I
think this is okay, since if there are 70-bit values there will probably
also be 71-bit values that will trigger the assert.