Prevent a process from attaching to a dead tracer.
When a tracer is terminated, it sends kill signals to all of its tracees.
This does not stop future tracees from being added to the terminated
tracer's list. Prevent this by checking the thread count of the tracer
before adding new tracees.

This was found by inspection and could only happen if a tracer's exit
raced with the initial debug enable on its child or while a descendant of
said child forks a new tracee process.