Tracing processes must wait for completion event.
In order to issue a debug command, a tracing process modifies its tracee's
debug command data structure and then signals an event to wake up the
tracee. It then needs to wait until the tracee signals the command
completion event before moving on. If it does not wait, the tracing
process may issue another command before the tracee has completed the
first command. This may cause the tracee to read a garbled message and
produce unintended consequences.

This was recently changed in order to fix a deadlock (see ba7f0985). That
deadlock occurs when a dying tracer sends a continue command to a waiting
tracee to make sure the tracee doesn't wait forever. It did this, however,
while holding its process queued lock. The issue is that the waiting child
may not yet be waiting. It may be trying to acquire the tracer's process
queued lock in order to signal that its about to wait. The deadlock is the
tracee waiting on the tracer's process queued lock and the tracer waiting
for the tracee to signal the command completion event. The fix was to
allow a continue, signle step, and range step command to exit without
waiting for the completion event. Besides the issue outlined in the
paragraph above, this cause asserts to fire during range step operations
as the tracee's command structure was cleaned by the exiting tracer.

So, the fix involves treating the tracee/tracer relationship more like the
child/parent process relationship. This allows a dying tracer to send that
continue command outside of its process queued lock. This means that a
tracee no longer holds a reference on its tracer, requiring all tracees to
acquire their process lock, check to see if the tracer process pointer is
not NULL and grabbing a reference to proceed using the tracer process
pointer outside the lock.