Add Broadcom 27xx I2C controller.
This is untested on Raspberry Pi devices, the driver is not included in any
images, and the firmware does not enumerate the three Broadcom Serial
Controller I2c busses. We're in need of an I2C device to test it on, but it
has been sitting on my machine for a while and I wanted to push it before
it got lost. It will hopefully work with minimal debugging effort.