Bump automation page file size to 3GB.
Boren (x86q) is failing to build groff due to the regex that blows up
badly. There is still lots of room in the page file, but apparently not
enough contiguous room for a very large section. Try a larger page file,
otherwise the regex rework project will become a bit more urgent.`