Added shmget, shmat, shmdt, and shmctl for mesa.
This change adds the older shmget style shared memory interface, which is
used somewhat by Mesa. My approach was to basically add it to the existing
shared memory support, so all shared memory objects are capable of acting
as both the older SystemV style and POSIX style shared memory objects.

I had to bend the OS mmap functions a bit so that if given a size of zero
they would figure out the actual size of the object at hand.