Disallow mmapping of directories.
mmapping of directories didn't work anyway, because the mapping would
succeed, but then the first access to it would cause a fault. Freetype was
dying because it thought it was good after successfully mmapping a
directory (as told to by mkfontscale).

This change fails mmaps of directories in the map call itself, preventing
the insanity.