Add FIONREAD (SIOCINQ) support for stream sockets.
This change adds the missing FIONREAD ioctl for stream (TCP) sockets. It
reports the amount of data available to be read in the socket's receive
buffer.

Other systems have support for this ioctl for datagram (UDP) sockets, but
as it only reports the size of the next datagram, it's better to just poll
datagram sockets to know when data arrives. But it would be easy to add
support later if some application finds this ioctl necessary for UDP.