Preserve ignored signals across exec.
The kernel was correctly preserving ignored signals across exec, the
problem is the C library structures thought everything was set to default.
So the behavior is all correct, but the return values coming from
signal/sigaction were wrong.

Turns out xinit sets SIGUSR1 to SIG_IGN before launching Xorg (the X
server). Xorg is supposed to see SIGUSR1 is ignored, and take that as an
indication to send its parent a SIGUSR1 once the X server is fully
initialized. Xorg was calling signal/sigaction, but getting back a return
of SIG_DFL. So it never signaled xinit. This wasn't fatal, but caused
xinit to delay 15 seconds before firing up the desktop, which was
apparently annoying enough to debug.