Better handoff of the framebuffer to user mode.
This change improves the kernel removing its hands from the framebuffer
when user mode opens it. The videoconsole continues to read from the
master terminal device, but no longer forwards that information into the
screen.

Also added a system information call to be able to enable and disable the
banner thread. The video console uses this to disable the banner thread
when the frame buffer is opened, and potentially re-enable it and redraw
that area once closed.

Also removed the old periodic reboot stuff.