Fix assert for active TCP sockets.
When a TCP packet arrived, the tree of fully bound sockets was searched for
an active socket that matched the address tuple. Once the socket's lock was
acquired and packet processing began, TCP asserted that the socket was
still active. This assert was, however, a little too high up. It was
asserting that sockets not in the initialized state were active, but closed
sockets can also be inactive. This change just moves the assert below
packet processing for closed sockets.