fread should read everything or have an excuse.
According to POSIX, the only reason fread should return less than the
desired amount is EOF or error. We were only performing a single read in
the case of non-buffered streams, so we were violating that. Python's
subprocess module was bailing out early because it thought it was done
reading.

This change reads in a loop to get all data. Interestingly fwrite was
already doing this.