Fix for regexes with funky characters.
The file utility was trying to compile a regular expression that contained
something like [\x8a] (where the converted the \x into an actual 0x8A
byte. We had missed a couple casts to UCHARs, so this got upconverted to a
signed int, and then converted to a ULONG of 0xFFFFFF8A, which failed
because that's a crazy value for a byte.