Add device context to user control IRP.
A user I/O request is performed on a specific handle, but the device
driver's context for that handle was not being passed down in the user
control IRP. Thus far, user control IRPs were only making device wide
changes. This change adds it to the IRP structure so that a device can
make handle specific changes.