Fix for spurious error prints on Windows swiss.
SwOpen on Windows was calling SwpPrintLastError on every open that failed.
This is an issue because 1) not all failed opens should be reported to the
user, and 2) Most that are were already being reported by the applet. The
tricky part is that CreateFile, used to get O_APPEND correct, doesn't set
errno, which is what all the applets look at when something went wrong.
The solution is to call the regular open() on the way out so that errno
gets set. If that open happens to succeed, the file descriptor is
returned, although it might not be fully O_APPEND if that happens to
occur.