Allow the STATIC_TLS flag if it has no TLS image.
Static TLS during dlopen is normally prohibited because there's only a
limited amount of space allocated for the initial TLS allocation, and the
initial-exec model of TLS just reaches directly to an offset from the
thread pointer.

swrast_dri.so (part of mesa of course) has the STATIC_TLS flag set, but
doesn't actually have a TLS image. Relax the restriction to allow this,
since it must just be doing TLS reference to variables in other libraries.