Move X packages into main native compile.
It would have been nice to encapsulate the X packages into their own
tasks. The trouble is the previous working directory would be removed and
a new one added for the new task. When rebuilding or upgrading the
automation system, consider a different scheme for managing the working
directory (perhaps per-schedule rather than per-task, or perhaps with a
way of keeping the working directory for nested tasks).