Add memory-based interface for lzma.
This change adds support for a memory buffer based interface to the LZMA
encoder/decoder, in addition to the Read()/Write() interface that's
already there. A combination of use is also possible (ie filling a read
buffer but supplying a Write() function).

I think the memory buffer interface is more natural for a lot of use cases
than the Read()/Write() functions. It works better in cases where you're
trying to simply insert LZMA into a file stream like interface, where you
may not have all the input data ready to be streamed in at the outset.

With this change the encoder works, which is the hard part. The decoder is
broken and I haven't debugged that yet. It should be much easier because
the decoder basically worked in this manner anyway.