Changes that shook out of trying -flto.
LTO isn't actually turned on anywhere, but I was playing with building
with it turned on. Mostly we seem to be okay on the optimization front.
The biggest problem is that the compiler doesn't recognize that functions
called by assembly are in fact used, so it optimizes them away unless
they're explicitly marked as used.

The other thing we were assuming is that if we specified an object file
first, it would end up at the front of the binary. I had to move a bunch
of assembly to the .init section in order to keep it at the front of the
binary with -flto turned on. It did allow me to remove an old custom
PandaBoard linker script though, so that's always a plus.

The one weirdness I couldn't track down is intrinsic functions like
__divdi3 and __aeabi_uidiv seemed to be undefined references in the linker
with flto turned on during their compilation, even if they were explicitly
marked as used. I never did figure that one out, and ended up turning off
lto for that Rtl base library just to get around it.