Better protect ATA channels from each other.
ATA has a single set of registers to speak to one of two devices on a
cable (called a task file or channel). We weren't properly synchronizing
access to the two different devices on the same cable. The rule is that if
you're calling AtapSelectDevice and touching registers, you need to be
holding the shared controller DPC lock. If you're not, then the other
device on the cable could select the device away from you, and your
register reads/writes would go to the wrong device.

This might explain some of David's hangs/interrupt storms.