Remove paging in asserts of size == pagesize.
This change relaxes the asserts that say when paging in that the first
I/O buffer fragment better be a page long (as in, we better have read in a
full page). This is not true for framebuffer devices that don't end on a
page boundary.

This means that usermode can map a full page even if the I/O completed
wasn't a full page. For things like the frame buffer, this means they'll
get access to the remainder of that page. This seems necessary in order to
map hardware directly. Hopefully this doesn't mean leaking access to other
resources or uninitialized kernel memory.