Add the sound core library driver.
This driver can partner with any sound controller driver to do a lot of
the work it takes to support sound. It handles OSS-compatible ioctls,
I/O buffer production/consumption, and even helps enumerate the devices
attached to the controller with a miniature pseudo-FS.

User mode can use OsLocateDeviceInformation to find devices that support
the SOUND_DEVICE_INFORMATION_UUID, call OsOpenDevice to open the sound
controller directory and then use OsOpen with an "output" or "input" path
to open a device relative to the controller directory.