Propagate map flags from file objects to mmap.
X on ARM wasn't looking so hot because it had cached mappings of the frame
buffer. It's a pretty interesting looking effect, but not super useful.

This change adds a MapFlags member to the Lookup IRP, which allows a file
object to turn on additional map flags. This lives in the file object, and
propagates out to image sections and I/O buffer mappings of page cache
entries in that file object. Private mappings of those file objects keep
their usual flags, since it's always a private copy for the process.