Chalk: Better executable location awareness.
Based on argv[0], chalk is now able to find its absolute executable path.
It's true that argv[0] could be set to anything, so as of now doing
something misleading with argv[0] is unsupported. If Chalk is needed for
suid environments, more robust but OS-specific methods of getting the
executable path can be explored, like getexecpath.

With the newly found executable path, Chalk also adds
dirname(execpath)/../lib/chalk to the library path. This way Chalk can
find its libraries even if it's relocated or in a non-standard location.
Quite nice.