Wire mingen into build, and build.ck fixes.
mingen is now another app that gets built. Also fixed several build.ck
failures that accumulated along the way.

Also added a complaint in the currently active minoca.mk build system if
the current directory is not a subdirectory of SRCROOT. This happens if
there are symlinks somewhere in the path, since make goes and gets the
physical path.