Improved stack unwinding for !thread.
This change does two things. The first thing it does for !thread is to
scoop up all the registers pushed by the assembly code that it can, giving
it a better chance at unwinding correctly.

The second change is to unwind the register defined as the CFA using the
CFA rule. I couldn't find anywhere in the DWARF specification where it
says to do this, but the problem I'm seeing is that no rules are being
defined for unwinding the stack register. So the frames higher up the
stack are using the same SP as the original frame. I'm guessing you're
supposed to unwind that register to the CFA value. Right now it's in a
conditional so that it's only 1) if the CFA is register based (obvi) and
2) if that register was left undefined otherwise.

What we really need are some .cfi directives in a few key places
throughout the assembly, like in the exception code and syscall code. That
would allow us to successfully unwind through things like interrupts and
page faults.