Added infrastructure for bootstrap Makefiles.
There's a circular problem with the new mingen-based build system: the
sources to build mingen are in this tree, but the future build
infrastructure will need mingen as part of the build process. So you need
mingen to build mingen.

The solution I'm thinking is basically to check in a minimal subset of the
output of a previous run of mingen. Someone with a fresh checkout will use
the bootstrap makefiles to build mingen, and then can use that mingen to
create the complete set of build files (for all build OSes).

Since mingen targets the build machine, the Makefiles produced will be
different for each different build OS. The hope is that most of the less
mainstream OSes will be able to use Minoca's build file, as it's fairly
vanilla Unixy. Porting to a new build OS will involve either manually
tweaking the first bootstrap Makefile, or generating one from an OS that's
already supported.

In order to keep the bootstrap Makefiles from going stale, their
regeneration is wired into the main build process. This means that the
source tree is written to during builds. Unless something changed, the
output should be identical, so source control shouldn't complain unless
the file really needs to be updated.

I needed to add a few options to mingen to allow me to create the same
set of bootstrap Makefiles every time. This means no random
/some/user/location paths floating around.

Also fixed a couple issues with the new build system building the
BeagleBone Black and PandaBoard firmware.