Add mmap support to sound core.
This change allows a sound device to be memory mapped. This is similar to
how the frame buffer is mmapped - if an I/O buffer with no fragments is
supplied on a read, the sound device's internal buffer is appended to the
empty read buffer.

This adds the generic MmAppendIoBuffer routine that will append one I/O
buffer (starting at an offset) to the end of another I/O buffer.