Add support for the OSS set trigger ioctl.
The set trigger ioctl allows a user to manually enable/disable a sound
device. Thus, it is renamed internally to SoundEnableDevice. If the device
is set to be disabled, then it will not automatically start on the first
read/write call. If it is then manually enabled later, it will immediately
start playing/recording sound.