Add sound ioctls to query hardware position.
OSS defines ioctls that allow the user to determine where the hardware is
operating within its buffer. These are only useful when the user has
mmaped the sound device.

These ioctls also report the total number of bytes processed by the device
since it was opened/reset. In order to support this the sound core API that
allows a sound controller to update the buffer's I/O state needed to be
more generic so that sound core could determine how far the controller
moved forward. It now takes and sets the controller's offset, rather than
assuming the controller had set it.