More frame unwinder bug fixes.
1. Stop looking for stack loops, since the detection method of the frame
pointer not changing might not be valid for architectures like ARM,
where moving the stack pointer isn't technically required for a
function call.

2. If the return register is specified, make the default rule for that
register be 'same value'. Breaking in at the beginning of an ARM function
doesn't have any rules for lr, yet the return register is valid so the
frame can be unwound.

3. Don't update unwound registers until all registers have been unwound.
They might depend on each other (eg. ebp is at [esi+4], but esi has
already been restored, so the restoration of ebp is all wrong).

4. Don't unconditionally set the return address in the "PC" register.
Well, actually do, but do it before the other registers have been written,
so that if there is an explicit restoration for the PC register it wins.
We were having trouble with our trap frame unwinding because we want to
restore both LR and PC, but we get confused if we just smash the return
address register (LR) on top of PC. Then there's no way to restore them to
different values.

5. Fix a typo where if any register had an expression rule we'd set it
as the rule for the CFA register.

6. Dwarf DW_CFA_def_cfa_expression should use the resulting value as a CFA
directly, not as a memory address to dereference. I found the
documentation confusing because they make this distinction between
DW_CFA_expression, annotated with expression(E), and
DW_CFA_val_expression, annotated with val_expression(E).
DW_CFA_def_cfa_expression is annotated with expression(E), so you'd think
it would need a dereference, but it doesn't, and the wording in the
paragraph vaguely agrees that it doesn't.