Add .cfi annotations to assembly files.
This change adds C preprocessor macros in the assembly that expand out
into .cfi_* directives, and sprinkles those directives around a few key
functions. I have no interest in going crazy and annotating everything,
since the annotations are distracting and difficult to maintain. But
annotating a few functions that you're likely to find on call stacks is
very handy. It's very nice to be able to see through a page fault,
interrupt, system call, and signal.