Enforce a minimum sound buffer fragment count of 2.
User mode may specify 0 or 1 fragments in the buffer size hint ioctl. The
OSS documentation for the SNDCTL_DSP_SETFRAGMENT ioctl says the fragment
count should be a minimum of 2. The user may be ignoring that and
providing a bad value. If 0 or 1 is supplied, use the sound controllers
minimum number of required fragments.