Pass the signal context up through the C library.
The sa_sigaction form of a signal function takes a signal context as its
third parameter. Gperftools is using that to get the PC from within a
signal. We were always passing NULL.

This change percolates that context up from libminocaos through the C
library and to the user's signal handler.