Allow a driver to request a non-paged I/O state.
The Intel HD Audio driver needs to signal the sound device's I/O state in
a DPC to make sure user mode has enough time to read/write the DMA buffer.
Waiting for a work item to signal the event can be too slow if all the
cores are near the beginning of their quantum when the interrupt fires.

This change adds a new lookup flag, that allows a device to specify that
it needs a non-paged I/O state.