Move Intel HDA stream processing to DPC.
Under heavier workloads (e.g. X running fceux), it was taking an average
of 7ms and a maximum of 68ms to get from the ISR to the work item. This
was ruining the audio quality. For Mario, SDL only had 1024 byte fragments
and for 16-bit mono audio at 44.1Khz, each fragment took 11.6ms to play.
If SDL doesn't get updated until 7ms into the 11.6ms, not much time is
left to get to the data for the next frame prepared. Moving to the DPC
leaves plenty of time.

Under light workloads, the ISR to work item was taking an average of 14
microseconds (max of 750 microseconds). So, the system isn't slow, it's
just busy and the worker thread gets delayed.