Bump Intel HDA output buffer offset by the FIFO size.
An Intel HD Audio controller interrupts for an output stream's fragment
when the last byte of data has been fetched from memory and put into the
DMA FIFO. This means that the interrupt DPC will likely read back an offset
that is a FIFO size short of a fragment boundary. Sound core (and user mode)
really want to know when the fragment is complete, so innocently skip the
offset ahead by a FIFO size.

In the future, this may not work out for video/audio synchronization. If
sound core and user mode really need to know when the fragment has been
played, then some high priority work item will need be scheduled to run
shortly after the DPC runs to double check the buffer position.

This is not applicable to input streams, as those do not interrupt until
all of the data has been transfered from the DMA FIFO into the descriptor's
buffer.