Always signal on sound controller buffer updates.
If the sound controller is calling into sound core to update its buffer
offset, assume that more data is now available (bytes to read or empty
space to write into). As such, always signal the event. This eliminates
a corner case in IoSysPoll where the polling thread is woken up but then
doesn't see any events. This was happening when running fceux while
toggling xterm focus in X. Fceux's sound thread would poll on the sound
device and be signaled. But by the time it woke up, the sound controller
hardware had made a full pass on the buffer and unsignled the event (it
saw the offsets as equal and thought there was no space available).