Don't assume all sound controllers support MMAP.
Sound core used to always tack on the MMAP and manual enable capabilities,
but some sound controllers (e.g. RPi's PWM Audio) cannot support mmap as
the sound data in PCM format needs to be translated to PWM format. In an
mmap scenario, there is no good time to perform this translation.