DMA changes for Raspberry Pi PWM Audio.
The DMA transfer for PWM audio needs to loop continuously once submitted.
It should only stop when canceled, but needs to interrupt periodically
after each audio fragment and loop back to the first fragment once the
last completes.

In order to support this, a generic DMA_TRANSFER_CONTINUOUS flag was added.
A controller can advertise support for this flag via its DMA information
capabilities when registering with the DMA core. Paired with this is an
InterruptPeriod size, which specifies after how many bytes a continuous DMA
transfer should interrupt.

Another fix was needed in the Broadcom DMA ISR, as it was clearing a DMA
channels active bit when meaning to only clear the interrupt bit.