Add optional copy buffer callback to sound core.
In case a sound controller wants to handle copies into or out of the sound
devices buffer, this change allows for an optional callback. This is
needed for the Raspberry Pi PWM audio as the PWM audio driver needs to
convert the sound data from PCM format to PWM format. As PWM audio format
is not a real format, sound core and user mode do not need to know about
it to play audio on the Raspberry Pi.

This change also makes sure copy sizes are aligned to the sample size.