Add PWM initialization to RPi firmware.
This change initializes PWM for audio output. It sets GPIO pins 40 and 45
to ALT0 for the left (PWM0) and right (PWM1) audio channels of the
headphone jack. It also initializes the PWM clock to a reasonable rate
(100MHz) and passes this clock value to the kernel via the Broadcom ACPI
table.