Add PWM audio driver for Raspberry Pi devices.
This change adds support for Raspberry Pi audio using the headphone jack.
The jack is wired up to GPIO pins 40 and 45, which in turn are connected
to PWM0 and PWM1. With a simple conversion between 16-bit PCM data and
32-bit PWM values, the audio actually isn't too bad.

On Raspberry Pi Model B+ and later, the audio channels are actually
inverted. They switched GPIO pin 40 from left to right and GPIO pin 45 from
right to left. The Raspberry Pi firmware made a change to swap these back
in software, but there is no documentation as to how this is done. That's
the next step.