Add low water threshold ioctl to sound core.
OSS offers an ioctl that dictates how many bytes need to be available
before a sound device is signalled. This change implements support for
that.

To make this easier, as a bit of clean-up, sound buffers now keep track
of the number of available bytes, rather than having to calculate them
based on the core and controller offsets. This actually ends up making
a few other ioctls and the I/O routine simpler. And to make some of the
calculations performant, buffer size, fragment size, and fragment count
must be a power of 2.