Add a small test.sh for lzma library.
This change adds a test I had been working on a few weeks ago for the lzma
library. It grabs a couple big binaries from third-party, compresses them
at a couple different levels and in a couple different styles,
decompresses them, and compares against the original.

I'd like this test to compare the compressed CRCs against known-good
checksums created by the oracle, so that if a bug is introduced into
the encoder that causes suboptimal but still correct encoding it will be
spotted.