Ignore garbage when getting i8042 keyboard response.
VirtualBox was failing keyboard initialization if the user typed anything
in earlier because those keycodes were coming in while we were sending the
reset command. I'm not sure if this is a VirtualBox bug or if we aren't
properly disabling scanning in the keyboard. With this loop, scancodes are
ignored, and eventually either the request times out or a recognized
response is received.