Add basic realm management to Santa.
This change adds a RealmManager class that globally manages all active
realms. The RealmManager is responsible for creating, enumerating, and
destroying realms, as well as loading and saving their configurations to
disk. The Realm class is an instance of a realm. It's responsible for
setting up sharing between a realm and the root, and ties together the
three data layout tuples: Containment, Storage, and Presentation.

Instances of the Container class manage world isolation. Currently chroot
and none are implemented, and cgroups would be managed by another subclass
of Containment.

The Storage class defines package data storage. This is used to define
where package contents should be extracted to. Currently basic (a
directory store) and none (extract directly to final location) are
implemented. One could also imagine an romount type, which is similar to
basic except the presentation layer would be pointed at a read-only bind
mount of the store, so that changes cannot be accidentally propagated
between realms.

The Presentation class defines how a file is placed into its final
location. Currently only "copy" is implemented, where the file is copied
from the store to its final location. Future types might be "hardlink",
where the files are hardlinked from the store, "symlink", where the files
are symlinked to the store, and "overlay", where overlayfs mounts are
placed to create the destination.

I've also added basic commands to exercise the RealmManager: new-realm for
creating a realm, del-realm for deleting one, and list-realms for
enumerating the existing realms. Right now I'm thinking the policy is that
realms that begin with underscore are reserved, so when a package is being
built it can create a realm without fear of name collision with a realm
the user has created.